+++ controller/Changelog.md
@@ -3,6 +3,7 @@
 ## Changed
 
 - os: Update nixpkgs channel to 21.11
+- os: Disable virtual terminals that are not used by PlayOS
 
 ## Fixed